Factored lemmas from field_verus into field_lemmas
