+++ curve25519-dalek/src/backend/serial/u64/field_verus.rs
@@ -16,8 +16,10 @@ use super::common_verus::shift_lemmas::*;
 
 use super::field_lemmas::as_nat_lemmas::*;
 use super::field_lemmas::field_core::*;
+use super::field_lemmas::negate_lemmas::*;
 use super::field_lemmas::pow2_51_lemmas::*;
 use super::field_lemmas::pow2k_lemmas::*;
+use super::field_lemmas::reduce_lemmas::*;
 
 // ADAPTED CODE LINES: X.0 globally replaced with X.limbs
 
@@ -27,255 +29,14 @@ verus! {
 // LOW_51_BIT_MASK: u64 = (1u64 << 51) -1; originally
 pub const LOW_51_BIT_MASK: u64 = 2251799813685247u64; // 2^51  -1
 
-// Basic properties of LOW_51_BIT_MASK:
-// - It's the value of low_bits_mask (spec function defined in vstd and used in its lemmas)
-// - it's less than 2^51
-pub proof fn l51_bit_mask_lt()
+// Establishes LOW_51_BIT_MASK == mask51 and lifts properties of the RHS to the LHS
+pub proof fn masks_are_equal()
     ensures
         LOW_51_BIT_MASK == mask51,
         LOW_51_BIT_MASK == low_bits_mask(51),
         LOW_51_BIT_MASK < (1u64 << 51) as nat,
 {
-    lemma2_to64_rest();
-    assert(LOW_51_BIT_MASK < (1u64 << 51) as nat) by (compute);
-}
-
-// Masking with low_bits_mask(51) gives a value bounded by 2^51
-pub proof fn masked_lt_51(v: u64)
-    ensures
-        v & LOW_51_BIT_MASK < (1u64 << 51),
-{
-    l51_bit_mask_lt(); // LOW_51_BIT_MASK == low_bits_mask(51)
-    masked_lt(v, 51);
-}
-
-// lemma_div_and_mod specialization for k = 51, using LOW_51_BIT_MASK == low_bits_mask(51)
-pub proof fn lemma_div_and_mod_51(ai:u64, bi: u64, v: u64)
-    requires
-        ai == v >> 51,
-        bi == v & LOW_51_BIT_MASK
-    ensures
-        ai == v / (pow2(51) as u64),
-        bi == v % (pow2(51) as u64),
-        v == ai * pow2(51) + bi
-{
-    l51_bit_mask_lt(); // LOW_51_BIT_MASK == low_bits_mask(51)
-    lemma_div_and_mod(ai, bi, v, 51);
-}
-
-pub broadcast proof fn lemma_cast_then_mask_51(x: u128)
-    ensures
-        #![trigger (x as u64) & LOW_51_BIT_MASK]
-        (x as u64) & LOW_51_BIT_MASK == x & (LOW_51_BIT_MASK as u128)
-{
-    assert((x as u64) & 2251799813685247u64 == x & (2251799813685247u64 as u128)) by (bit_vector);
-}
-
-pub open spec fn spec_reduce(limbs: [u64; 5]) -> (r: [u64; 5]) {
-    let r = [
-        ((limbs[0] & LOW_51_BIT_MASK) + (limbs[4] >> 51) * 19) as u64,
-        ((limbs[1] & LOW_51_BIT_MASK) + (limbs[0] >> 51)) as u64,
-        ((limbs[2] & LOW_51_BIT_MASK) + (limbs[1] >> 51)) as u64,
-        ((limbs[3] & LOW_51_BIT_MASK) + (limbs[2] >> 51)) as u64,
-        ((limbs[4] & LOW_51_BIT_MASK) + (limbs[3] >> 51)) as u64,
-    ];
-    r
-}
-
-// Each component of spec_reduce is bounded.
-// The reason we _don't_ write
-// ensures forall |i: int| 0 <= i < 5 ==> spec_reduce(limbs)[i] < (1u64 << 52)
-// is that the solver treats `spec_reduce`` above as symbolic and does _not_ instantiate e.g.
-// ((limbs[4] & LOW_51_BIT_MASK) + (limbs[3] >> 51)) as u64 < (1u64 << 52)
-pub proof fn lemma_boundaries(limbs: [u64; 5])
-    ensures
-        ((limbs[0] & LOW_51_BIT_MASK) + (limbs[4] >> 51) * 19) < (1u64 << 52),
-        ((limbs[1] & LOW_51_BIT_MASK) + (limbs[0] >> 51)) < (1u64 << 52),
-        ((limbs[2] & LOW_51_BIT_MASK) + (limbs[1] >> 51)) < (1u64 << 52),
-        ((limbs[3] & LOW_51_BIT_MASK) + (limbs[2] >> 51)) < (1u64 << 52),
-        ((limbs[4] & LOW_51_BIT_MASK) + (limbs[3] >> 51)) < (1u64 << 52)
-
-{
-    // \A i. limbs[i] < 2^13
-    shifted_lt(limbs[0], 51);
-    shifted_lt(limbs[1], 51);
-    shifted_lt(limbs[2], 51);
-    shifted_lt(limbs[3], 51);
-    shifted_lt(limbs[4], 51);
-
-    // \A i. limbs[i] & LOW_51_BIT_MASK < 2^51
-    masked_lt_51(limbs[0]);
-    masked_lt_51(limbs[1]);
-    masked_lt_51(limbs[2]);
-    masked_lt_51(limbs[3]);
-    masked_lt_51(limbs[4]);
-
-    // Since 19 < 2^5 and (limbs[4] >> 51) < 2^13, their product is less than 2^18
-    assert((limbs[4] >> 51) * 19 < (1u64 << 18) as nat) by {
-        assert(19 < (1u64 << 5)) by (bit_vector);
-        shift_is_pow2(5);
-        shift_is_pow2(13);
-        shift_is_pow2(18);
-        lemma_pow2_adds(13, 5);
-        // If (limbs[4] >> 51) < 2^13 and 19 < 2^5 then their product is less than 2^18
-        mul_lt((limbs[4] >> 51) as nat, (1u64 << 13) as nat, 19nat, (1u64 << 5) as nat);
-    }
-
-    // The final values (limbs[i] += cX) are all bounded by 2^51 + eps, for eps \in {2^18, 2^13}.
-    assert(((1u64 << 18)) + (1u64 << 51) < (1u64 << 52)) by (bit_vector);
-    assert(((1u64 << 13)) + (1u64 << 51) < (1u64 << 52)) by (bit_vector);
-
-    // In summary, they're all bounded by 2^52
-    // The solver can prove this automatically
-}
-
-pub proof fn lemma_reduce(limbs: [u64; 5])
-    ensures
-        forall|i: int| 0 <= i < 5 ==> spec_reduce(limbs)[i] < (1u64 << 52),
-        // Suppose l = (l0, l1, l2, l3, l4) are the input limbs.
-        // They represent a number
-        // e(l) =  l0 + l1 * 2^51 + l2 * 2^102 + l3 * 2^153 + l4 * 2^204
-        // in Z_p, for p = 2^255 - 19
-        // reduce(l) returns v = (v0, v1, v2, v3, v4), such that
-        // v0 = 19 * a4 + b0
-        // v1 =      a0 + b1
-        // v2 =      a1 + b2
-        // v3 =      a2 + b3
-        // v4 =      a3 + b4
-        // where ai = li >> 51 and bi = li & LOW_51_BIT_MASK
-        // we can reformulate this as ai = li / 2^51 (flooring division) and bi = li % 2^51
-        // Using the following identity connecting integer division and remainder:
-        // x = y * (x / y) + x % y
-        // we can see that li = ai * 2^51 + bi
-        // Plugging the above identities into the equations for v, we can observe that
-        // e(v) = e(l) - p * (l4 >> 51)
-        // IOW, e(reduce(l)) = e(l) (mod p)
-        // additionally, if all limbs are below 2^51, reduce(l) = l
-        (forall|i: int| 0 <= i < 5 ==> limbs[i] < (1u64 << 51)) ==> (spec_reduce(limbs) =~= limbs),
-        as_nat(spec_reduce(limbs)) == as_nat(limbs) - p() * (limbs[4] >> 51)
-{
-
-    // -----
-    // reduce identity for small limbs
-
-    // Can't seem to reference r within this proof block, we reconstruct it here
-    let rr: [u64; 5] = spec_reduce(limbs);
-
-    assert((forall|i: int| 0 <= i < 5 ==> #[trigger] limbs[i] < (1u64 << 51)) ==> (rr =~= limbs)) by {
-        if (forall|i: int| 0 <= i < 5 ==> #[trigger] limbs[i] < (1u64 << 51)) {
-            assert forall|i: int| 0 <= i < 5 implies #[trigger] limbs[i] & LOW_51_BIT_MASK == limbs[i] by {
-                l51_bit_mask_lt(); // LOW_51_BIT_MASK = low_bits_mask(51)
-                shift_is_pow2(51);
-                lemma_u64_low_bits_mask_is_mod(limbs[i], 51);
-                lemma_small_mod(limbs[i] as nat, pow2(51));
-            }
-            assert forall|i: int| 0 <= i < 5 implies #[trigger] limbs[i] >> 51 == 0 by {
-                l51_bit_mask_lt(); // LOW_51_BIT_MASK = low_bits_mask(51)
-                shift_is_pow2(51);
-                lemma_u64_shr_is_div(limbs[i], 51);
-                lemma_basic_div(limbs[i] as int, pow2(51) as int);
-            }
-        }
-    }
-
-    // -- as_nat identity
-
-    // ai = limbs[i] / 2^52
-    let a0 = (limbs[0] >> 51);
-    let a1 = (limbs[1] >> 51);
-    let a2 = (limbs[2] >> 51);
-    let a3 = (limbs[3] >> 51);
-    let a4 = (limbs[4] >> 51);
-
-    // bi = limbs[i] % 2^52
-    let b0 = (limbs[0] & LOW_51_BIT_MASK);
-    let b1 = (limbs[1] & LOW_51_BIT_MASK);
-    let b2 = (limbs[2] & LOW_51_BIT_MASK);
-    let b3 = (limbs[3] & LOW_51_BIT_MASK);
-    let b4 = (limbs[4] & LOW_51_BIT_MASK);
-
-    lemma_boundaries(limbs);
-
-    // distribute
-    assert(as_nat(rr) ==
-        19 *  a4 + b0 +
-        pow2(51) * a0 + pow2(51) * b1 +
-        pow2(102) * a1 + pow2(102) * b2 +
-        pow2(153) * a2 + pow2(153) * b3 +
-        pow2(204) * a3 + pow2(204) * b4
-    ) by {
-        lemma_mul_is_distributive_add(pow2( 51) as int, a0 as int, b1 as int);
-        lemma_mul_is_distributive_add(pow2(102) as int, a1 as int, b2 as int);
-        lemma_mul_is_distributive_add(pow2(153) as int, a2 as int, b3 as int);
-        lemma_mul_is_distributive_add(pow2(204) as int, a3 as int, b4 as int);
-    }
-
-    // factor out
-    assert(as_nat(rr) ==
-        19 *  a4 + b0 +
-        pow2(51) * a0 + pow2(51) * b1 +
-        pow2(51) * (pow2(51) * a1) + pow2(102) * b2 +
-        pow2(102) * (pow2(51) * a2) + pow2(153) * b3 +
-        pow2(153) * (pow2(51) * a3) + pow2(204) * b4
-    ) by {
-        lemma_two_factoring_51(51, a1);
-        lemma_two_factoring_51(102, a2);
-        lemma_two_factoring_51(153, a3);
-    }
-
-    // change groupings
-    assert(as_nat(rr) ==
-        (b0 + pow2(51) * a0) +
-        pow2(51) * (b1 + pow2(51) * a1) +
-        pow2(102) * (b2 + pow2(51) * a2) +
-        pow2(153) * (b3 + pow2(51) * a3) +
-        pow2(204) * b4 + 19 * a4
-    ) by {
-        lemma_mul_is_distributive_add(pow2( 51) as int, b1 as int, pow2(51) * a1);
-        lemma_mul_is_distributive_add(pow2(102) as int, b2 as int, pow2(51) * a2);
-        lemma_mul_is_distributive_add(pow2(153) as int, b3 as int, pow2(51) * a3);
-    }
-
-    // invoke div/mod identity
-    assert(as_nat(rr) ==
-        limbs[0] +
-        pow2(51) * limbs[1] +
-        pow2(102) * limbs[2] +
-        pow2(153) * limbs[3] +
-        pow2(204) * b4 + 19 * a4
-    ) by {
-        lemma_div_and_mod_51(a0, b0, limbs[0]);
-        lemma_div_and_mod_51(a1, b1, limbs[1]);
-        lemma_div_and_mod_51(a2, b2, limbs[2]);
-        lemma_div_and_mod_51(a3, b3, limbs[3]);
-    }
-
-    // Add missing limbs[4] parts
-    assert(as_nat(rr) ==
-        limbs[0] +
-        pow2(51) * limbs[1] +
-        pow2(102) * limbs[2] +
-        pow2(153) * limbs[3] +
-        pow2(204) * limbs[4] - pow2(204) * (pow2(51) * a4 ) + 19 * a4
-    ) by {
-        lemma_div_and_mod_51(a4, b4, limbs[4]);
-        assert(pow2(204) * limbs[4] == pow2(204) * b4 + pow2(204)* (pow2(51) * a4)) by {
-            lemma_mul_is_distributive_add(pow2(204) as int, pow2(51) * a4 as int, b4 as int);
-        }
-    }
-
-    // The solver can collect components of as_nat(limbs) automatically:
-    // as_nat(rr) == as_nat(limbs) - pow2(204) * (pow2(51) * a4 ) + 19 * a4
-    // ... as well as pull in minus
-    // as_nat(rr) == as_nat(limbs) - (pow2(204) * (pow2(51) * a4 ) - 19 * a4)
-
-    // collect components of p() * a4
-    assert(pow2(204) * (pow2(51) * a4) - 19 * a4 == p() * a4) by {
-        lemma_mul_is_associative(pow2(204) as int, pow2(51) as int, a4 as int);
-        lemma_pow2_adds(204, 51);
-        lemma_mul_is_distributive_sub_other_way(a4 as int, pow2(255) as int, 19 );
-        pow255_gt_19(); // we need to prove 2^255 - 19 doesn't underflow
-    }
+    l51_bit_mask_lt();
 }
 
 /* MANUALLY moved outside, named return value */
@@ -359,84 +120,8 @@ impl FieldElement51 {
             (as_nat(self.limbs) + as_nat(old(self).limbs)) % p() == 0
     {
         proof {
-            let c0 = (pow2(51) - 19);
-            let c  = (pow2(51) - 1);
-            lemma2_to64_rest(); // get pow2(51)
-            // solver knows 36028797018963664u64 == 16 * c0
-            // solver knows 36028797018963952u64 == 16 * c;
-
-            assert forall |i: int| 0 <= i < 5 implies old(self).limbs[i] < 16 * c0 by {
-                shift_is_pow2(51);
-            }
-
-            // Introduce 16p as a vector
-            let v = [(16 * c0) as u64,(16 * c) as u64,(16 * c) as u64,(16 * c) as u64,(16 * c) as u64];
-
-            assert(as_nat(v) == 16 * p()) by {
-                // by definition of as_nat
-                assert( as_nat(v) ==
-                    16 * c0 +
-                    pow2(51) * (16 * c) +
-                    pow2(102) * (16 * c) +
-                    pow2(153) * (16 * c) +
-                    pow2(204) * (16 * c)
-                );
-
-                // solver can reorder factors and pull out 16 on its own
-                // ...
-
-                // Write out `c`s and sum up powers
-                assert( p() ==
-                    c0 +
-                    pow2(51) * c +
-                    pow2(102) * c +
-                    pow2(153) * c +
-                    pow2(204) * c
-                ) by {
-                    lemma_pow2_adds(51, 51);
-                    lemma_pow2_adds(51, 102);
-                    lemma_pow2_adds(51, 153);
-                    lemma_pow2_adds(51, 204);
-                }
-            }
-
-            let l0 = old(self).limbs[0];
-            let l1 = old(self).limbs[1];
-            let l2 = old(self).limbs[2];
-            let l3 = old(self).limbs[3];
-            let l4 = old(self).limbs[4];
-
-            assert(as_nat([
-                (16 * c0 - l0) as u64,
-                (16 * c - l1) as u64,
-                (16 * c - l2) as u64,
-                (16 * c - l3) as u64,
-                (16 * c - l4) as u64,
-                ]) == as_nat(v) - as_nat(old(self).limbs)
-            ) by {
-                lemma_as_nat_sub(v, old(self).limbs);
-            }
-
-            let k = (16 * c - l4) as u64 >> 51;
-
-            assert(
-                16 * p() - as_nat(old(self).limbs) - p() * k + as_nat(old(self).limbs)
-                ==
-                p() * (16 - k)
-            ) by {
-                lemma_mul_is_distributive_sub(p() as int, 16, k as int)
-            }
-
-            assert((p() * (16 - k)) as nat % p() == 0) by {
-                assert(k <= 16) by {
-                    assert(k <= (16 * pow2(51)) as u64 >> 51) by {
-                        lemma_shr_le_u64((16 * c - l4) as u64, (16 * pow2(51)) as u64, 51);
-                    }
-                    // 16 * 2^51 / 2^51 = 16
-                    assert(((16 * 0x8000000000000) as u64 >> 51) == 16) by (compute);
-                }
-                lemma_mod_multiples_basic((16 - k) as int, p() as int);
-            }
+            lemma_neg_no_underflow(self.limbs);
+            negate_proof(self.limbs);
         }
         // See commentary in the Sub impl: (copied below)
             // To avoid underflow, first add a multiple of p.
